requirements.txt
rich
cryptography
pyOpenSSL
questionary
tabulate
requests 